Update dot_map.py
New aproach to algorithm.
New approach to the algorithm
Until now, the algorithm calculated the BBOX (bounding box) of each polygon and generated random points based on those limits. If a point was contained within the polygon (contains), it was considered a candidate. 
The performance problem occurred when the geometry was a multipolygon formed by several isolated (non-contiguous) polygons, such as an archipelago of islands. The probability of a point falling outside the polygon is very high, which meant that the calculation time was excessively long.
We now apply a new approach to the problem. If the geometry is a multipolygon with the characteristics indicated above, the following is done:
1. The number of polygons is counted.
2. The BBOX of each polygon is calculated. 
2. The area of each polygon is stored.
3. A number of points is assigned to each polygon in proportion to its area. (based on Hamilton's method for seat allocation).
4. For each polygon, the corresponding points are generated based on its BBOX.
This significantly reduces the calculation time.